Simplified, renamed, documented example program
