+++ wdtk-examples/src/resources/sample-dump-20150815.json.gz
